Update new_collect_file.py
In the previous verion of the script I had used "content.txt" instead of "collect.txt"  Which is not compatible with the main script. 

In this version, I have corrected that error, and I have also created a method by which you can change the name of the txt file using a --fix command.